Fixed bug in return value of rift functions

+++ rift.h
@@ -13,7 +13,7 @@
 
 // ********************* Interface      ********************* //
 
-uint8_t rift_send_keep_alive( hid_device *handle );
-uint8_t rift_send_enable_components( hid_device *handle, uint8_t on_u8 );
+int16_t rift_send_keep_alive( hid_device *handle );
+int16_t rift_send_enable_components( hid_device *handle, uint8_t on_u8 );
 
 #endif
